Adapterの作成と結果リスト表示確認
